requirements.txt
click
Flask==2.0.2
itsdangerous
Jinja2==3.0.3
MarkupSafe
pip
setuptools
Werkzeug==2.0.2
Flask-SQLAlchemy
Flask-Login
pyotp
pycryptodome
SQLAlchemy
WTForms==3.0.1
cryptography
Flask-WTF==1.0.0
email-validator==1.1.3
pymysql
sshtunnel==0.4.0

